requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.9
aiosignal==1.3.1
analytics-python==1.4.post1
annotated-types==0.7.0
anyio==4.6.2.post1
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==24.2.0
audioread==3.0.1
aws-cdk-lib==2.171.1
aws-cdk.asset-awscli-v1==2.2.213
aws-cdk.asset-kubectl-v20==2.1.3
aws-cdk.asset-node-proxy-agent-v6==2.1.0
aws-cdk.cloud-assembly-schema==38.0.1
backoff==1.10.0
bcrypt==4.2.1
beautifulsoup4==4.12.3
bleach==6.2.0
blis==0.7.11
cachetools==5.5.0
catalogue==2.0.10
cattrs==24.1.2
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
cloudpathlib==0.19.0
colorama==0.4.6
comm==0.2.2
confection==0.1.5
constructs==10.4.2
contourpy==1.2.1
cryptography==43.0.3
cycler==0.12.1
cymem==2.0.8
cytoolz==0.12.3
debugpy==1.8.5
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==4.1.13
et-xmlfile==1.1.0
executing==2.0.1
fastapi==0.115.6
fastjsonschema==2.20.0
ffmpy==0.4.0
filelock==3.16.1
floret==0.10.5
fonttools==4.53.1
frozenlist==1.5.0
fsspec==2024.10.0
gradio==5.8.0
gradio_client==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.0
huggingface-hub==0.26.3
idna==3.10
imageio==2.35.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipykernel==6.29.5
ipython==8.26.0
jedi==0.19.1
jellyfish==1.1.0
Jinja2==3.1.4
joblib==1.4.2
jsii==1.105.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.2
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
kagglehub==0.3.2
kiwisolver==1.4.5
langcodes==3.4.1
language_data==1.2.0
lazy_loader==0.4
librosa==0.10.2.post1
linkify-it-py==2.0.3
llvmlite==0.43.0
marisa-trie==1.2.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mdurl==0.1.2
mistune==3.0.2
mlxtend==0.23.3
monotonic==1.6
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
murmurhash==1.0.10
nbclient==0.10.0
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.3
nltk==3.9.1
numba==0.60.0
numpy==2.1.3
openpyxl==3.1.5
orjson==3.10.12
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pandoc==2.4
pandocfilters==1.5.1
paramiko==3.5.0
parso==0.8.4
patsy==0.5.6
pdf2image==1.17.0
pdfminer.six==20231228
pdfplumber==0.11.4
pillow==11.0.0
platformdirs==4.2.2
plumbum==1.9.0
ply==3.11
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.47
propcache==0.2.1
psutil==6.0.0
psycopg2==2.9.9
publication==0.0.3
pure_eval==0.2.3
pycparser==2.22
pycryptodome==3.21.0
pydantic==2.10.3
pydantic_core==2.27.1
pydub==0.25.1
pyee==11.1.1
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.4
PyPDF2==3.0.1
pypdfium2==4.30.0
pyphen==0.16.0
pyppeteer==2.0.0
PySocks==1.7.1
pytesseract==0.3.13
pythainlp==5.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.19
pytube==15.0.0
pytz==2024.2
pywin32==306
PyYAML==6.0.2
pyzmq==26.2.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
ruff==0.8.1
safehttpx==0.1.6
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
seaborn==0.13.2
selenium==4.27.1
semantic-version==2.10.0
sentencepiece==0.2.0
setuptools==75.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.0.4
sniffio==1.3.1
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.5.0.post1
spacy==3.7.6
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlparse==0.5.1
srsly==2.4.8
stack-data==0.6.3
starlette==0.41.3
statsmodels==0.14.2
sympy==1.13.1
textacy==0.13.0
thinc==8.2.5
threadpoolctl==3.5.0
tifffile==2024.9.20
tinycss2==1.4.0
tokenizers==0.20.3
tomlkit==0.13.2
toolz==0.12.1
torch==2.5.1
tornado==6.4.1
tqdm==4.67.1
traitlets==5.14.3
transformers==4.46.3
trio==0.27.0
trio-websocket==0.11.1
typeguard==2.13.3
typer==0.15.1
typing_extensions==4.12.2
tzdata==2024.2
uc-micro-py==1.0.3
urllib3==2.2.3
uvicorn==0.32.1
virtualenv==20.28.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webdriver-manager==4.0.2
webencodings==0.5.1
websocket-client==1.8.0
websockets==14.1
wordcloud==1.9.3
wrapt==1.16.0
wsproto==1.2.0
xgboost==2.1.2
xlrd==2.0.1
yarl==1.18.3
zipp==3.20.2
